% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pm25}
\alias{example_pm25}
\title{Example timeseries dataset}
\format{
A data frame with 2 columns:
\describe{
  \item{datetime}{POSIXct timestamp of the observation}
  \item{pm25}{PM2.5 concentration (µg/m³)}
}
}
\usage{
example_pm25
}
\description{
The `example_pm25` dataset provides example timeseries data for practicing
and code examples. This dataset represents hourly air quality measurements.

The dataset was generated on 2021-09-22 by running:

\preformatted{
library(AirSensor)

example_pm25 <- example_sensor$data
names(example_pm25) <- c("datetime", "pm25")

save(example_pm25, file = "data/example_pm25.rda")
}
}
\keyword{datasets}
