% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaRollUtils-package.R
\docType{package}
\name{MazamaRollUtils-package}
\alias{MazamaRollUtils}
\alias{MazamaRollUtils-package}
\title{MazamaRollUtils: Efficient Rolling Functions}
\description{
Fast rolling-window functions for numeric vectors. Designed for efficient processing of environmental time-series data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MazamaScience/MazamaRollUtils}
  \item \url{https://mazamascience.github.io/MazamaRollUtils/}
  \item Report bugs at \url{https://github.com/MazamaScience/MazamaRollUtils/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Callahan \email{jonathan.s.callahan@gmail.com}

Authors:
\itemize{
  \item Hans Martin \email{hansmrtn@gmail.com}
}

}
\keyword{internal}
