% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NowCast.R
\name{nowcast}
\alias{nowcast}
\title{Calculate NowCast}
\usage{
nowcast(x, n = 12, threshold = 0.5)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{

}
\description{
As described
\href{https://usepa.servicenowservices.com/airnow?id=kb_article_view&sys_id=fed0037b1b62545040a1a7dbe54bcbd4}{here}.
}
