\name{MazamaRollUtils-package}
\alias{MazamaRollUtils-package}
\alias{MazamaRollUtils}
\docType{package}
\title{
  Mazama Science Rolling Utilities
}
\description{
  A suite of utility functions for calculating rolling mins, means,
  maxes and other functions written with an efficient Rcpp/C++ backend.
}
\author{
  Jonathan Callahan,
  Hans Martin
}

\keyword{ package }
