\name{MazamaRollUtils-package}
\alias{MazamaRollUtils-package}
\alias{MazamaRollUtils}
\docType{package}
\title{
  This package implements a number of rolling functions for R.
}
\description{
  A suite of utility functions for calculating rolling mins, means,
  maxes and other functions with a fast C++ backend implementation.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Jonathan Callahan,
Hans Martin

Maintainer: Hans Martin <hans@mazamascience.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
