% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_max}
\alias{roll_max}
\title{Roll Max}
\usage{
roll_max(x, n = 5L, weights = NULL, by = 1L, align = 0L)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Integer window size.}

\item{weights}{Numeric vector of length \code{n} specifying each window
index weight. If \code{NULL}, the unit weight is used.}

\item{by}{Integer shift to use when sliding the window to the next location}

\item{align}{Signed integer representing the window alignment.
\code{-1(left)|0(center)|1(right)}.}
}
\value{
numeric vector of length(x)
}
\description{
Apply a moving-window maximum function to a numeric vector.
}
\details{
Each window of \code{n}-length is applied \code{weight} and then
slid/shifted/rolled \code{by} a positive integer amount about the window's
\code{align}-ment index.

The \code{align} parameter determines the alignment of the current index
within the window. Thus:

\itemize{
  \item{\code{align = -1 [*------]} will cause the returned vector to have n-1 \code{NA} values at the right end.}
  \item{\code{align = 0 [---*---]} will cause the returned vector to have (n-1)/2 \code{NA} values at either end.}
  \item{\code{align = 1 [------*]} will cause the returned vector to have n-1 \code{NA} values at the left end.}
}
}
\examples{
# Load package air quality data
data("airquality")

# Calculate moving maximum of adjacent measurements
roll_mean(airquality$Temp, n = 3)
}
