% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_max}
\alias{roll_max}
\title{Roll Max}
\usage{
roll_max(x, n = 5L, weights = NULL, by = 1L, align = 0L)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer window length.}

\item{weights}{A numeric vector of n-length specifying each \code{n}th
weight. If `NULL`, unit weights are used.}

\item{by}{An integer to shift the window by.}

\item{align}{A signed integer representing the windows alignment.
-1 (left) | 0 (center) | 1 (right).}
}
\value{
numeric vector of length(x)
}
\description{
Roll Max
}
\examples{
# load airquality
data("airquality")

# calculate moving maximum of adjacent measurements
roll_mean(airquality$Temp, n = 3)
}
