% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_max}
\alias{roll_max}
\title{Roll Max}
\usage{
roll_max(x, width = 5L, by = 1L, align = 0L)
}
\arguments{
\item{x}{Numeric vector.}

\item{width}{Integer width of the rolling window.}

\item{by}{Integer shift to use when sliding the window to the next location}

\item{align}{Signed integer representing the position of the return value within each window.
\code{-1(left) | 0(center) | 1(right)}.}
}
\value{
Numeric vector of the same length as \code{x}.
}
\description{
Apply a moving-window maximum function to a numeric vector.
}
\details{
For every index in the incoming vector \code{x}, a value is returned that
is the maximum of all values in \code{x} that fall within a window of width
\code{width}.

The \code{align} parameter determines the alignment of the current index
within the window. Thus:

\itemize{
  \item{\code{align = -1 [*------]} will cause the returned vector to have width-1 \code{NA} values at the right end.}
  \item{\code{align = 0  [---*---]} will cause the returned vector to have width/2 \code{NA} values at either end.}
  \item{\code{align = 1  [------*]} will cause the returned vector to have width-1 \code{NA} values at the left end.}
}

For large vectors, the\code{by} parameter can be used to force the window
to jump ahead \code{by} indices for the next calculation. Indices that are
skipped over will be assigned \code{NA} values so that the return vector still has
the same length as the incoming vector. This can dramatically speed up
calculations for high resolution time series data.
}
\examples{
library(MazamaRollUtils)

# R default air quality time series
t <- datasets::airquality$Temp

plot(t)
lines(roll_max(t, width = 5), col = 'salmon')
}
