% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roll.R
\name{roll_mean}
\alias{roll_mean}
\title{Roll Mean}
\usage{
roll_mean(x, n = 5L, by = 1L, align = 0L)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{An integer window size}

\item{by}{Integer to shift window by}

\item{align}{Signed Integer window alignment, -1 (left) | 0 (center) | 1 (right)}
}
\value{
numeric vector of length(X)
}
\description{
Roll Mean
}
