% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roll.R
\name{roll_mean}
\alias{roll_mean}
\title{Roll Mean}
\usage{
roll_mean(x, n = 5L, weights = NULL, by = 1L, align = 0L)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{An integer window size #'}

\item{weights}{A numeric vector of n-length specifying each \code{n}th weight.}

\item{by}{Integer to shift window by}

\item{align}{Signed Integer window alignment, -1 (left) | 0 (center) | 1 (right)}
}
\value{
numeric vector of length(x)
}
\description{
Roll Mean
}
