% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_min}
\alias{roll_min}
\title{Roll Min}
\usage{
roll_min(x, n = 5L, weights = NULL, by = 1L, align = 0L)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer window length.}

\item{weights}{A numeric vector of size \code{n} specifying each window
index weight. If \code{NULL}, the unit weight is used.}

\item{by}{An integer to shift the window by.}

\item{align}{A signed integer representing the windows alignment.
\code{-1(left)|0(center)|1(right)}.}
}
\value{
numeric vector of length(x)
}
\description{
Apply a moving-window minimum function to to a numeric vector.
}
\details{
Each window of \code{n}-length is applied \code{weight} and then
slid/shifted/rolled \code{by} a positive integer amount about the window's
\code{align}-ment index.
}
\examples{
# load airquality
data("airquality")

# calculate moving minimum of last 24 measurements
roll_min(airquality$Temp, n = 24, align = -1)
}
