% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaRollUtils.R
\name{roll_sd}
\alias{roll_sd}
\title{Roll Standard Deviation}
\usage{
roll_sd(x, width = 1L, by = 1L, align = c("center", "left", "right"))
}
\arguments{
\item{x}{Numeric vector.}

\item{width}{Integer width of the rolling window.}

\item{by}{Integer shift by which the window is moved each iteration.}

\item{align}{Character position of the return value within the window. One of:
`"left" | "center" | "right"`.}
}
\value{
Numeric vector of the same length as `x`.
}
\description{
Apply a moving-window standard deviation function to a
numeric vector.
}
\details{
For every index in the incoming vector `x`, a value is returned that
is the standard deviation of all values in `x` that fall within a window of
width `width`.

The `align` parameter determines the alignment of the return value
within the window. Thus:

\itemize{
  \item{`align = "left"   [*------]` will cause the returned vector to have width - 1 `NA` values at the right end.}
  \item{`align = "center" [---*---]` will cause the returned vector to have `NA` values at either end as needed for centered alignment.}
  \item{`align = "right"  [------*]` will cause the returned vector to have width - 1 `NA` values at the left end.}
}

For large vectors, the `by` parameter can be used to force the window
to jump ahead `by` indices for the next calculation. Indices that are
skipped over will be assigned `NA` values so that the return vector still has
the same length as the incoming vector. This can dramatically speed up
calculations for high resolution time series data.
}
\note{
A `na.rm` argument is not provided for `roll_sd()` because the
statistical meaning of standard deviation computed from partially missing
windows may be ambiguous.
}
\examples{
# Example air quality time series
t <- example_pm25$datetime
x <- example_pm25$pm25

x[1:10]
roll_sd(x, width = 5)[1:10]
}
