% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_sum}
\alias{roll_sum}
\title{Roll Sum}
\usage{
roll_sum(x, width = 5L, weights = NULL, by = 1L, align = 0L)
}
\arguments{
\item{x}{Numeric vector.}

\item{width}{Integer width of the rolling window.}

\item{weights}{A numeric vector of size \code{n} specifying each window
index weight. If \code{NULL}, the unit weight is used.}

\item{by}{An integer to shift the window by.}

\item{align}{A signed integer representing the windows alignment.
\code{-1(left)|0(center)|1(right)}.}
}
\value{
numeric vector of length(x)
}
\description{
Apply a moving-window sum to a numeric vector.
}
\details{
Each window of \code{n}-length is applied \code{weight} and then
slid/shifted/rolled \code{by} a positive integer amount about the window's
\code{align}-ment index.

The \code{align} parameter determines the alignment of the current index
within the window. Thus:

\itemize{
  \item{\code{align = -1 [*------]} will cause the returned vector to have width-1 \code{NA} values at the right end.}
  \item{\code{align = 0  [---*---]} will cause the returned vector to have width/2 \code{NA} values at either end.}
  \item{\code{align = 1  [------*]} will cause the returned vector to have width-1 \code{NA} values at the left end.}
}
}
\examples{
# load airquality
data("airquality")

# calculate moving sum of last 3 measurements
roll_sum(airquality$Temp, width = 3, align = -1)
}
