% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_sum}
\alias{roll_sum}
\title{Roll Sum}
\usage{
roll_sum(x, n = 5L, weights = NULL, by = 1L, align = 0L)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer window length.}

\item{weights}{A numeric vector of n-length specifying each \code{n}th
weight. If `NULL`, unit weights are used.}

\item{by}{An integer to shift the window by.}

\item{align}{A signed integer representing the windows alignment.
-1 (left) | 0 (center) | 1 (right).}
}
\value{
numeric vector of length(x)
}
\description{
Roll Sum
}
\examples{
# load airquality
data("airquality")

# calculate moving sum of last 3 measurements
roll_sum(airquality$Temp, n = 3, align = -1)
}
