% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_var}
\alias{roll_var}
\title{Roll Variance}
\usage{
roll_var(x, width = 5L, by = 1L, align = 0L)
}
\arguments{
\item{x}{Numeric vector.}

\item{width}{Integer width of the rolling window.}

\item{by}{An integer to shift the window by.}

\item{align}{A signed integer representing the position of the return value within each window.
\code{-1(left)|0(center)|1(right)}.}
}
\value{
Numeric vector of the same length as \code{x}.
}
\description{
Apply a moving-window variance function to a numeric vector.
}
\details{
Each window of \code{n}-length is applied \code{weight} and then
slid/shifted/rolled \code{by} a positive integer amount about the window's
\code{align}-ment index.

The \code{align} parameter determines the alignment of the current index
within the window. Thus:

\itemize{
  \item{\code{align = -1 [*------]} will cause the returned vector to have width-1 \code{NA} values at the right end.}
  \item{\code{align = 0  [---*---]} will cause the returned vector to have width/2 \code{NA} values at either end.}
  \item{\code{align = 1  [------*]} will cause the returned vector to have width-1 \code{NA} values at the left end.}
}
}
\examples{
# load airquality
data("airquality")

# calculate moving variance of adjacent measurements
roll_mean(airquality$Temp, width = 3)
}
